requirements.txt
requests
pyyaml
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
streamlit
fastapi
uvicorn
